+++ packages/react-magma-dom/src/components/Combobox/Combobox.stories.tsx
@@ -19,10 +19,8 @@ export default {
   component: Combobox,
   argTypes: {
     labelPosition: {
-      control: {
-        type: 'select',
-        options: LabelPosition,
-      },
+      control: { type: 'select' },
+      options: Object.values(LabelPosition),
     },
     labelWidth: {
       control: {