+++ packages/react-magma-dom/src/components/Search/Search.stories.tsx
@@ -31,10 +31,8 @@ export default {
       },
     },
     inputSize: {
-      control: {
-        type: 'select',
-        options: InputSize,
-      },
+      control: { type: 'select' },
+      options: Object.values(InputSize),
     },
     isPredictive: {
       defaultValue: false,