+++ packages/react-magma-dom/src/components/TimePicker/TimePicker.stories.tsx
@@ -28,10 +28,8 @@ export default {
       },
     },
     labelPosition: {
-      control: {
-        type: 'select',
-        options: LabelPosition,
-      },
+      control: { type: 'select' },
+      options: Object.values(LabelPosition),
     },
     labelWidth: {
       control: {