+++ website/react-magma-docs/CHANGELOG.md
@@ -1,5 +1,26 @@
 # Change Log
 
+## 7.0.0-rc.8
+
+### Patch Changes
+
+- ccb95ae: fix(NavTabs, Tabs, Toast): Fix different issues.
+- d473442: fix(Select): fix issue when scrollbar appears because of padding and negative margin.
+- Updated dependencies [ed37531]
+  - react-magma-dom@5.1.0-rc.13
+
+## 7.0.0-rc.7
+
+### Patch Changes
+
+- 998a3e3: chore: Update React-magma-icons version to 3.2.2
+
+## 7.0.0-rc.6
+
+### Patch Changes
+
+- ca15e5a: fix(Storybook): Fix issue with `type: 'select'`.
+
 ## 7.0.0-rc.5
 
 ### Patch Changes